requirements.txt
PyYAML
requests
pandas
numpy
xlrd
TA-Lib
tables
schedule
wxpusher
pytest
baostock
akshare
DingtalkChatbot
baostock
tqdm